OBEU-28390: Change types

+++ src/main/java/com/icekome/inditex/application/PriceFinder.java
@@ -8,8 +8,8 @@ public interface PriceFinder {
 
   List<Price> findBy(
       LocalDate date,
-      int productList,
-      int chainId
+      long productList,
+      long chainId
   );
 
 }
